Extract MaskPattern to another file
